Integrate `generate_macro_def_id_path` into `href_with_root_path`
